ekisde

+++ AlohAndes/src/uniandes/isis2304/alohandes/persistencia/SQLHospedaje.java
@@ -5,36 +5,39 @@ import java.util.List;
 
 import javax.jdo.PersistenceManager;
 import javax.jdo.Query;
-
-import uniandes.isis2304.alohandes.negocio.Habitacion.TipoHabitacion;
 import uniandes.isis2304.alohandes.negocio.Hospedaje;
 
 public class SQLHospedaje
 {
 	private final static String SQL = PersistenciaAlohandes.SQL;
-	
+
 	private PersistenciaAlohandes pp;
-	
+
 	public SQLHospedaje (PersistenciaAlohandes pp)
 	{
 		this.pp = pp;
 	}
-	
+
 	public long adicionarHospedaje (PersistenceManager pm, long id, String nombre, String direccion, double valorTotal) 
 	{
-        Query q = pm.newQuery(SQL, "INSERT INTO " + pp.darTablaHospedaje() + "((idHospedaje , nombre, direccion, valorTotal ) values (?, ?, ?, ?)");
-        q.setParameters(id, nombre, direccion, valorTotal);
-        return (long) q.executeUnique();
+		Query q = pm.newQuery(SQL, "INSERT INTO " + pp.darTablaHospedaje() + "((idHospedaje , nombre, direccion, valorTotal ) values (?, ?, ?, ?)");
+		q.setParameters(id, nombre, direccion, valorTotal);
+		return (long) q.executeUnique();
 	}
-	
+
 
 	public long eliminarHospedajePorId (PersistenceManager pm, long idHospedaje)
 	{
-        Query q = pm.newQuery(SQL, "DELETE FROM " + pp.darTablaHospedaje () + " WHERE idHospedaje = ?");
-        q.setParameters(idHospedaje);
-        return (long) q.executeUnique();
+		SQLContrato c = new SQLContrato(pp);
+		if(c.darContratosPorIdHospedaje(pm, idHospedaje).isEmpty())
+		{
+			Query q = pm.newQuery(SQL, "DELETE FROM " + pp.darTablaHospedaje () + " WHERE idHospedaje = ?");
+			q.setParameters(idHospedaje);
+			return (long) q.executeUnique();
+		}
+		return -1;
 	}
-	
+
 	public Hospedaje darHospedajePorId (PersistenceManager pm, long idHospedaje) 
 	{
 		Query q = pm.newQuery(SQL, "SELECT * FROM " + pp.darTablaHospedaje () + " WHERE idHospedaje = ?");
@@ -42,7 +45,7 @@ public class SQLHospedaje
 		q.setParameters(idHospedaje);
 		return (Hospedaje) q.executeUnique();
 	}
-	
+
 	public List<Hospedaje> darHospedaje (PersistenceManager pm)
 	{
 		Query q = pm.newQuery(SQL, "SELECT * FROM " + pp.darTablaHospedaje ());
